change: 時間割コンポーネントのファイル分け
