+++ app/src/main/java/com/example/taskmate/navigation/Navigation.kt
@@ -13,10 +13,10 @@ import androidx.navigation.compose.NavHost
 import androidx.navigation.compose.composable
 import androidx.navigation.compose.currentBackStackEntryAsState
 import androidx.navigation.compose.rememberNavController
-import com.example.taskmate.ui.HomeScreen
 import com.example.taskmate.ui.MyPageScreen
 import com.example.taskmate.ui.SettingScreen
 import com.example.taskmate.ui.addTask.AddTaskScreen
+import com.example.taskmate.ui.home.HomeScreen
 import com.example.taskmate.ui.task.TaskScreen
 
 @Composable
