+++ app/src/main/java/com/example/taskmate/ui/home/components/WeekList.kt
@@ -1,0 +1,41 @@
+package com.example.taskmate.ui.home.components
+
+import androidx.compose.foundation.background
+import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.Box
+import androidx.compose.foundation.layout.Row
+import androidx.compose.foundation.layout.Spacer
+import androidx.compose.foundation.layout.height
+import androidx.compose.foundation.layout.width
+import androidx.compose.foundation.shape.RoundedCornerShape
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.unit.dp
+
+@Composable
+fun WeekList(dayOfWeekList: List<String>) {
+    Row(
+        horizontalArrangement = Arrangement.spacedBy(8.dp, Alignment.CenterHorizontally),
+        verticalAlignment = Alignment.CenterVertically,
+    ) {
+        Spacer(modifier = Modifier.width(50.dp))
+        dayOfWeekList.forEach { day ->
+            Box(
+                contentAlignment = Alignment.Center,
+                modifier =
+                Modifier
+                    .width(50.dp)
+                    .height(30.dp)
+                    .background(
+                        color = MaterialTheme.colorScheme.primary,
+                        shape = RoundedCornerShape(size = 5.dp),
+                    ),
+            ) {
+                Text(text = day, color = MaterialTheme.colorScheme.background)
+            }
+        }
+    }
+}
